Added a photos controller and upload_url to view stream.
Still is not working but I think is a good step in the right direction. We need to figure out how to add the image to the right stream. My idea is to use a hidden input field and see if that is preserved in the redirection once we get the post on BlobstoreUploadHandler.