Trending Cron,  Controller, &  Views
Added the cron job to generate the trending rank every 5 minutes and remove timestamps older than 1 hour.

Added Trend generator script, for which I had to add a view list, a view list count and a ranking to the Stream model

Hooked the calls to the trending views which displays the ordered list with view counts.

Added logic to add view timestamp when view single is triggered